+++ test/helpers/main.js
@@ -9,7 +9,7 @@ export const BaseError = ModernError.subclass('BaseError', {
 export const baseError = new BaseError('message')
 // eslint-disable-next-line fp/no-mutation
 baseError.one = true
-export const errorObject = baseError.toJSON()
+export const errorObject = BaseError.toJSON(baseError)
 
 export const nativeError = new TypeError('message')
 // eslint-disable-next-line fp/no-mutation
@@ -18,7 +18,7 @@ nativeError.one = true
 const parentNativeError = new BaseError('test')
 // eslint-disable-next-line fp/no-mutation
 parentNativeError.prop = nativeError
-export const nativeErrorObject = parentNativeError.toJSON().prop
+export const nativeErrorObject = BaseError.toJSON(parentNativeError).prop
 
 const testPlugin = {
   name: 'test',
@@ -34,4 +34,6 @@ export const PluginError = BaseError.subclass('PluginError', {
   plugins: [testPlugin],
 })
 export const pluginError = new PluginError('message', { test: true })
-export const pluginErrorObject = excludeKeys(pluginError.toJSON(), ['options'])
+export const pluginErrorObject = excludeKeys(PluginError.toJSON(pluginError), [
+  'options',
+])