+++ test/serialize.js
@@ -15,16 +15,46 @@ const convertError = function ({ name, message, stack, one }) {
   return { name, message, stack, one }
 }
 
-test('error.toJSON() serializes', (t) => {
+test('ErrorClass.toJSON() serializes', (t) => {
   t.deepEqual(errorObject, convertError(baseError))
 })
 
 each([baseError, nativeError], ({ title }, deepError) => {
-  test(`error.toJSON() is deep | ${title}`, (t) => {
+  test(`ErrorClass.toJSON() is deep | ${title}`, (t) => {
     const error = new BaseError('test')
     error.prop = [deepError]
-    t.deepEqual(error.toJSON().prop[0], convertError(deepError))
+    t.deepEqual(BaseError.toJSON(error).prop[0], convertError(deepError))
   })
+})
+
+test('ErrorClass.toJSON() keeps plugin options', (t) => {
+  t.true(pluginErrorObject.pluginsOpts.test)
+})
+
+test('ErrorClass.toJSON() keeps plugin options deeply', (t) => {
+  const error = new PluginError('message')
+  error.cause = new PluginError('causeMessage', { test: true })
+  t.true(PluginError.toJSON(error).cause.pluginsOpts.test)
+})
+
+test('ErrorClass.toJSON() does not keep plugin options of native errors', (t) => {
+  const error = new PluginError('message')
+  error.cause = new Error('causeMessage')
+  t.false('pluginsOpts' in PluginError.toJSON(error).cause)
+})
+
+test('ErrorClass.toJSON() does not keep empty plugin options', (t) => {
+  const error = new PluginError('message')
+  t.false('pluginsOpts' in PluginError.toJSON(error))
+})
+
+test('ErrorClass.toJSON() does not mutate error', (t) => {
+  PluginError.toJSON(pluginError)
+  t.false('pluginsOpts' in pluginError)
+})
+
+test('error.toJSON() serializes', (t) => {
+  t.deepEqual(baseError.toJSON(), convertError(baseError))
 })
 
 test('error.toJSON() is not enumerable', (t) => {
@@ -32,30 +62,4 @@ test('error.toJSON() is not enumerable', (t) => {
     Object.getOwnPropertyDescriptor(Object.getPrototypeOf(baseError), 'toJSON')
       .enumerable,
   )
-})
-
-test('error.toJSON() keeps plugin options', (t) => {
-  t.true(pluginErrorObject.pluginsOpts.test)
-})
-
-test('error.toJSON() keeps plugin options deeply', (t) => {
-  const error = new PluginError('message')
-  error.cause = new PluginError('causeMessage', { test: true })
-  t.true(error.toJSON().cause.pluginsOpts.test)
-})
-
-test('error.toJSON() does not keep plugin options of native errors', (t) => {
-  const error = new PluginError('message')
-  error.cause = new Error('causeMessage')
-  t.false('pluginsOpts' in error.toJSON().cause)
-})
-
-test('error.toJSON() does not keep empty plugin options', (t) => {
-  const error = new PluginError('message')
-  t.false('pluginsOpts' in error.toJSON())
-})
-
-test('error.toJSON() does not mutate error', (t) => {
-  pluginError.toJSON()
-  t.false('pluginsOpts' in pluginError)
 })